Separate C# v0 and v1 references
