+++ apps/docs/components/Navigation/NavigationMenu/HomeMenuIconPicker.tsx
@@ -26,7 +26,7 @@ import {
   IconMenuDevCli,
 } from './HomeMenuIcons'
 
-function getMenuIcon(menuKey: string, width: number = 16, height: number = 16, className) {
+function getMenuIcon(menuKey: string, width: number = 16, height: number = 16, className: string) {
   switch (menuKey) {
     case 'home':
       return <IconMenuHome width={width} height={height} className={className} />
