+++ apps/docs/components/Navigation/NavigationMenu/NavigationMenu.constants.ts
@@ -195,7 +195,7 @@ export const REFERENCES: References = {
   csharp: {
     name: 'C#',
     library: 'supabase-csharp',
-    versions: ['v0'],
+    versions: ['v1', 'v0'],
     icon: '/docs/img/libraries/c-sharp-icon.svg',
   },
   swift: {
@@ -2075,6 +2075,13 @@ export const reference_dart_v2 = {
 }
 
 export const reference_csharp_v0 = {
+  icon: 'reference-csharp',
+  title: 'C#',
+  url: 'guides/reference/csharp',
+  parent: '/reference',
+}
+
+export const reference_csharp_v1 = {
   icon: 'reference-csharp',
   title: 'C#',
   url: 'guides/reference/csharp',
