+++ apps/docs/components/Navigation/NavigationMenu/NavigationMenu.tsx
@@ -24,6 +24,7 @@ enum MenuId {
   RefDartV1 = 'reference_dart_v1',
   RefDartV2 = 'reference_dart_v2',
   RefCSharpV0 = 'reference_csharp_v0',
+  RefCSharpV1 = 'reference_csharp_v1',
   RefPythonV2 = 'reference_python_v2',
   RefSwiftV1 = 'reference_swift_v1',
   RefSwiftV2 = 'reference_swift_v2',
@@ -154,8 +155,16 @@ const menus: Menu[] = [
     commonSectionsFile: 'common-client-libs-sections.json',
     specFile: 'supabase_csharp_v0.yml',
     type: 'reference',
+    path: '/reference/csharp/v0',
+  },
+  {
+    id: MenuId.RefCSharpV1,
+    commonSectionsFile: 'common-client-libs-sections.json',
+    specFile: 'supabase_csharp_v1.yml',
+    type: 'reference',
     path: '/reference/csharp',
   },
+  ,
   {
     id: MenuId.RefPythonV2,
     commonSectionsFile: 'common-client-libs-sections.json',
