+++ apps/docs/docs/ref/csharp/v0/introduction.mdx
@@ -16,7 +16,7 @@ hideTitle: true
 
 {/* prettier-ignore */}
 <div className="max-w-xl">
-  This reference documents every object and method available in Supabase's C# library, [supabase](https://www.nuget.org/packages/supabase). You can use `Supabase` to interact with your Postgres database, listen to database changes, invoke Deno Edge Functions, build login and user management functionality, and manage large files.
+  This reference documents every object and method available in Supabase's C# library, [supabase-csharp](https://www.nuget.org/packages/supabase-csharp). You can use `Supabase` to interact with your Postgres database, listen to database changes, invoke Deno Edge Functions, build login and user management functionality, and manage large files.
 </div>
 
 {/* prettier-ignore */}
