+++ apps/docs/spec/common-client-libs-sections.json
@@ -32,6 +32,7 @@
       "reference_dart_v2",
       "reference_python_v2",
       "reference_csharp_v0",
+      "reference_csharp_v1",
       "reference_swift_v1",
       "reference_swift_v2",
       "reference_kotlin_v1",
@@ -48,6 +49,7 @@
       "reference_dart_v1",
       "reference_python_v2",
       "reference_csharp_v0",
+      "reference_csharp_v1",
       "reference_swift_v1",
       "reference_swift_v2",
       "reference_kotlin_v1",
@@ -986,7 +988,6 @@
       "reference_dart_v1",
       "reference_dart_v2",
       "reference_python_v2",
-      "reference_csharp_v0",
       "reference_swift_v1",
       "reference_swift_v2",
       "reference_kotlin_v1",
