+++ apps/docs/spec/supabase_csharp_v1.yml
@@ -9,11 +9,11 @@ info:
 
   definition: spec/enrichments/tsdoc_v2/combined.json
   slugPrefix: '/'
-  specUrl: https://github.com/supabase/supabase/edit/master/apps/docs/spec/supabase_csharp_v0.yml
+  specUrl: https://github.com/supabase/supabase/edit/master/apps/docs/spec/supabase_csharp_v1.yml
   libraries:
     - name: 'C#'
       id: 'csharp'
-      version: '0.8.0'
+      version: '1.0.0'
 
 functions:
   - id: initializing
